refactor: dedupe sheet-metal op + test boilerplate (SonarCloud duplication)
The Hem reuses the Flange's geometry but its operation/test wiring duplicated
the Flange's, tripping SonarCloud's new-code duplication gate (>3%). Extract
the shared scaffolding so each feature adds only its own substance:

- opregistry: activeSheetMetalPart(s, op) folds the ActivePart + sheet-metal
  guard the face/flange/hem operations all repeated into one helper.
- opregistry tests: seedSheetMetalSheet / topEdgeKey / expectMergedSolid shared
  across the flange and hem apply tests.
- model/feature tests: seedSheetMetalSheet builds the square wall + top edge the
  flange and hem geometry tests share, replacing five inline setups.

No behaviour change; all sheet-metal tests stay green.
